[OMNIML-2917] export layer config using actual prefix instead of hardcoded model.layers
This is change set 1 from working on OMNIML-2917.

When we export quantized model to hf unified format, we hard code
modules with a "model.layers" prefix. This is something completely odd
and unnecessary. The major problem is that we may output some quant
config that have completely wrong prefix, such as in exclude_modules.
For example, for the Qwen3-VL models, there are 2 transformer blocks:
language_model and vision. Before this change, for language_model, we
will output:

model.layers.language_model.layers.0.xxx
model.layers.language_model.layers.1.xxx

The prefixes are completely wrong therefore when inference systems such
as vllm try to read the quant config, it will fail.

Fix it by simply use the prefixes from parsing the model itself.

Signed-off-by: Shengliang Xu <shengliangx@nvidia.com>